backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.2.0
exifread==3.0.0
boto3==1.34.51
redis==5.0.2
xgboost==2.0.3
aiosqlite==0.20.0
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
websockets==12.0
httpx==0.27.0
pydantic-settings==2.2.1
hashlib2==1.0.0
